Added procfile

+++ index.js
@@ -13,13 +13,11 @@ client.on('message', msg => {
     const word = '*';
     const userID = process.env.ID;
 
-    if(msg.author.id === userID) {
-        if(msg.content.includes(word)) {
+    if(msg.author.id === userID && msg.content.includes(word)) {
             msg.delete();
             msg.channel.send(uwu(msg.content).split(word)); 
-        }
     }
     else return;
-});
+}); 
 
 client.login(process.env.TOKEN);
